+++ odoo/local-src/specific_zetes/wizard/__init__.py
@@ -1,1 +1,1 @@
-from . import change_lot
+from . import manage_uop